requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bottle==0.13.2
click==8.1.8
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
logger==1.4
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
proxy_tools==0.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyQt6-Qt6==6.8.2
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.2
PyQt6_sip==13.10.0
python-multipart==0.0.20
pywebview==5.4
QtPy==2.4.3
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
